Entity getting Stat fix
Main fix is using the property to get the attached `StatCollection`
instead of using the variable directly.  (do we really need to lazily
initialize?)

Also reduced the access level of the set for the Stat property to
protected so other people can't tell you what collection you're using.
(Do we even need a set?)